Compare the Consistency of Statistically Significant Features

This function compares the consistency and inconsistency of statistically significant features obtained using different methods in `pathway_daa` from the `ggpicrust2` package. It creates a report showing the number of common and different features identified by each method, and the features themselves.

Arguments

daa_results_list

A list of data frames containing statistically significant features obtained using different methods.

method_names

A character vector of names for each method used.

p_values_threshold

A numeric value representing the threshold for the p-values. Features with p-values less than this threshold are considered statistically significant. Default is 0.05.

Value

A data frame with the comparison results. The data frame has the following columns:

method: The name of the method.

num_features: The total number of statistically significant features obtained by the method.

num_common_features: The number of features that are common to other methods.

num_diff_features: The number of features that are different from other methods.

diff_features: The names of the features that are different from other methods.

Examples

# \donttest{
library(magrittr)
library(ggpicrust2)
library(tibble)
data("metacyc_abundance")
data("metadata")

# Run pathway_daa function for multiple methods
methods <- c("ALDEx2", "DESeq2", "edgeR","Maaslin2")
daa_results_list <- lapply(methods, function(method) {
pathway_daa(abundance = metacyc_abundance %>% column_to_rownames("pathway"),
metadata = metadata, group = "Environment", daa_method = method)
})
#> Sample names extracted.
#> Identifying matching columns in metadata...
#> Matching columns identified: sample_name . This is important for ensuring data consistency.
#> Using all columns in abundance.
#> Converting abundance to a matrix...
#> Reordering metadata...
#> Converting metadata to a matrix and data frame...
#> Extracting group information...
#> Running ALDEx2 with two groups. Performing t-test...
#> operating in serial mode
#> computing center with all features
#> ALDEx2 analysis with two groups complete.
#> Sample names extracted.
#> Identifying matching columns in metadata...
#> Matching columns identified: sample_name . This is important for ensuring data consistency.
#> Using all columns in abundance.
#> Converting abundance to a matrix...
#> Reordering metadata...
#> Converting metadata to a matrix and data frame...
#> Extracting group information...
#> Running DESeq2. Note: DESeq2 is only suitable for comparison between two groups.
#> Performing pairwise comparisons with DESeq2...
#> converting counts to integer mode
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> final dispersion estimates
#> fitting model and testing
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> -- replacing outliers and refitting for 12 genes
#> -- DESeq argument 'minReplicatesForReplace' = 7 
#> -- original counts are preserved in counts(dds)
#> estimating dispersions
#> fitting model and testing
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> Compiling DESeq2 results...
#> DESeq2 analysis complete.
#> Sample names extracted.
#> Identifying matching columns in metadata...
#> Matching columns identified: sample_name . This is important for ensuring data consistency.
#> Using all columns in abundance.
#> Converting abundance to a matrix...
#> Reordering metadata...
#> Converting metadata to a matrix and data frame...
#> Extracting group information...
#> Processing data with edgeR method...
#> Initializing edgeR object...
#> Calculating normalization factors...
#> Estimating common dispersions...
#> Disp = 0.69705 , BCV = 0.8349 
#> Performing exact test for two groups...
#> edgeR processing completed.
#> Sample names extracted.
#> Identifying matching columns in metadata...
#> Matching columns identified: sample_name . This is important for ensuring data consistency.
#> Using all columns in abundance.
#> Converting abundance to a matrix...
#> Reordering metadata...
#> Converting metadata to a matrix and data frame...
#> Extracting group information...
#> Running Maaslin2 analysis...
#> Performing Maaslin2 analysis...
#> 2023-09-09 09:30:49.741948 WARNING::Input is a matrix, passing through as.data.frame() .
#> [1] "Warning: Deleting existing log file: Maaslin2_results_Environment/maaslin2.log"
#> 2023-09-09 09:30:49.756657 INFO::Writing function arguments to log file
#> 2023-09-09 09:30:49.794479 INFO::Verifying options selected are valid
#> 2023-09-09 09:30:49.8581 INFO::Determining format of input files
#> 2023-09-09 09:30:49.858947 INFO::Input format is data samples as rows and metadata samples as rows
#> 2023-09-09 09:30:49.872411 INFO::Formula for fixed effects: expr ~  Environment
#> 2023-09-09 09:30:49.883147 INFO::Filter data based on min abundance and min prevalence
#> 2023-09-09 09:30:49.895906 INFO::Total samples in data: 50
#> 2023-09-09 09:30:49.899511 INFO::Min samples required with min abundance for a feature not to be filtered: 5.000000
#> 2023-09-09 09:30:49.915548 INFO::Total filtered features: 8
#> 2023-09-09 09:30:49.917146 INFO::Filtered feature names from abundance and prevalence filtering: GALLATE.DEGRADATION.I.PWY, METHYLGALLATE.DEGRADATION.PWY, P184.PWY, P281.PWY, PWY.1541, PWY.4361, PWY.5028, PWY.7527
#> 2023-09-09 09:30:49.943693 INFO::Total filtered features with variance filtering: 0
#> 2023-09-09 09:30:49.94601 INFO::Filtered feature names from variance filtering:
#> 2023-09-09 09:30:49.949842 INFO::Running selected normalization method: TSS
#> 2023-09-09 09:30:50.042484 INFO::Applying z-score to standardize continuous metadata
#> 2023-09-09 09:30:50.055275 INFO::Running selected transform method: AST
#> 2023-09-09 09:30:50.061253 INFO::Running selected analysis method: LM
#> 2023-09-09 09:30:50.067619 INFO::Fitting model to feature number 1, X1CMET2.PWY
#> 2023-09-09 09:30:50.076035 INFO::Fitting model to feature number 2, X3.HYDROXYPHENYLACETATE.DEGRADATION.PWY
#> 2023-09-09 09:30:50.081357 INFO::Fitting model to feature number 3, ALL.CHORISMATE.PWY
#> 2023-09-09 09:30:50.08409 INFO::Fitting model to feature number 4, ANAEROFRUCAT.PWY
#> 2023-09-09 09:30:50.086439 INFO::Fitting model to feature number 5, ANAGLYCOLYSIS.PWY
#> 2023-09-09 09:30:50.089185 INFO::Fitting model to feature number 6, ARG.POLYAMINE.SYN
#> 2023-09-09 09:30:50.092193 INFO::Fitting model to feature number 7, ARGDEG.PWY
#> 2023-09-09 09:30:50.095245 INFO::Fitting model to feature number 8, ARGORNPROST.PWY
#> 2023-09-09 09:30:50.098903 INFO::Fitting model to feature number 9, ARGSYN.PWY
#> 2023-09-09 09:30:50.108524 INFO::Fitting model to feature number 10, ARGSYNBSUB.PWY
#> 2023-09-09 09:30:50.113911 INFO::Fitting model to feature number 11, ARO.PWY
#> 2023-09-09 09:30:50.118934 INFO::Fitting model to feature number 12, ASPASN.PWY
#> 2023-09-09 09:30:50.128213 INFO::Fitting model to feature number 13, AST.PWY
#> 2023-09-09 09:30:50.131205 INFO::Fitting model to feature number 14, BIOTIN.BIOSYNTHESIS.PWY
#> 2023-09-09 09:30:50.133439 INFO::Fitting model to feature number 15, BRANCHED.CHAIN.AA.SYN.PWY
#> 2023-09-09 09:30:50.13673 INFO::Fitting model to feature number 16, CALVIN.PWY
#> 2023-09-09 09:30:50.13968 INFO::Fitting model to feature number 17, CENTFERM.PWY
#> 2023-09-09 09:30:50.142005 INFO::Fitting model to feature number 18, COA.PWY
#> 2023-09-09 09:30:50.144534 INFO::Fitting model to feature number 19, COBALSYN.PWY
#> 2023-09-09 09:30:50.146868 INFO::Fitting model to feature number 20, CODH.PWY
#> 2023-09-09 09:30:50.148982 INFO::Fitting model to feature number 21, COLANSYN.PWY
#> 2023-09-09 09:30:50.151236 INFO::Fitting model to feature number 22, COMPLETE.ARO.PWY
#> 2023-09-09 09:30:50.154371 INFO::Fitting model to feature number 23, DAPLYSINESYN.PWY
#> 2023-09-09 09:30:50.156978 INFO::Fitting model to feature number 24, DENOVOPURINE2.PWY
#> 2023-09-09 09:30:50.159398 INFO::Fitting model to feature number 25, DTDPRHAMSYN.PWY
#> 2023-09-09 09:30:50.161932 INFO::Fitting model to feature number 26, ECASYN.PWY
#> 2023-09-09 09:30:50.16432 INFO::Fitting model to feature number 27, ENTBACSYN.PWY
#> 2023-09-09 09:30:50.166625 INFO::Fitting model to feature number 28, FAO.PWY
#> 2023-09-09 09:30:50.169844 INFO::Fitting model to feature number 29, FASYN.ELONG.PWY
#> 2023-09-09 09:30:50.172787 INFO::Fitting model to feature number 30, FASYN.INITIAL.PWY
#> 2023-09-09 09:30:50.175344 INFO::Fitting model to feature number 31, FERMENTATION.PWY
#> 2023-09-09 09:30:50.177862 INFO::Fitting model to feature number 32, FOLSYN.PWY
#> 2023-09-09 09:30:50.18012 INFO::Fitting model to feature number 33, FUC.RHAMCAT.PWY
#> 2023-09-09 09:30:50.182251 INFO::Fitting model to feature number 34, FUCCAT.PWY
#> 2023-09-09 09:30:50.184568 INFO::Fitting model to feature number 35, GALACT.GLUCUROCAT.PWY
#> 2023-09-09 09:30:50.186803 INFO::Fitting model to feature number 36, GALACTARDEG.PWY
#> 2023-09-09 09:30:50.189014 INFO::Fitting model to feature number 37, GALACTUROCAT.PWY
#> 2023-09-09 09:30:50.192012 INFO::Fitting model to feature number 38, GLCMANNANAUT.PWY
#> 2023-09-09 09:30:50.194695 INFO::Fitting model to feature number 39, GLUCARDEG.PWY
#> 2023-09-09 09:30:50.196839 INFO::Fitting model to feature number 40, GLUCARGALACTSUPER.PWY
#> 2023-09-09 09:30:50.198929 INFO::Fitting model to feature number 41, GLUCONEO.PWY
#> 2023-09-09 09:30:50.20115 INFO::Fitting model to feature number 42, GLUCOSE1PMETAB.PWY
#> 2023-09-09 09:30:50.20329 INFO::Fitting model to feature number 43, GLUCUROCAT.PWY
#> 2023-09-09 09:30:50.205879 INFO::Fitting model to feature number 44, GLUTORN.PWY
#> 2023-09-09 09:30:50.209213 INFO::Fitting model to feature number 45, GLYCOCAT.PWY
#> 2023-09-09 09:30:50.211837 INFO::Fitting model to feature number 46, GLYCOGENSYNTH.PWY
#> 2023-09-09 09:30:50.214041 INFO::Fitting model to feature number 47, GLYCOL.GLYOXDEG.PWY
#> 2023-09-09 09:30:50.216763 INFO::Fitting model to feature number 48, GLYCOLYSIS
#> 2023-09-09 09:30:50.21999 INFO::Fitting model to feature number 49, GLYCOLYSIS.E.D
#> 2023-09-09 09:30:50.23208 INFO::Fitting model to feature number 50, GLYCOLYSIS.TCA.GLYOX.BYPASS
#> 2023-09-09 09:30:50.23495 INFO::Fitting model to feature number 51, GLYOXYLATE.BYPASS
#> 2023-09-09 09:30:50.237688 INFO::Fitting model to feature number 52, GOLPDLCAT.PWY
#> 2023-09-09 09:30:50.242113 INFO::Fitting model to feature number 53, HCAMHPDEG.PWY
#> 2023-09-09 09:30:50.245048 INFO::Fitting model to feature number 54, HEME.BIOSYNTHESIS.II
#> 2023-09-09 09:30:50.247554 INFO::Fitting model to feature number 55, HEMESYN2.PWY
#> 2023-09-09 09:30:50.249889 INFO::Fitting model to feature number 56, HEXITOLDEGSUPER.PWY
#> 2023-09-09 09:30:50.252132 INFO::Fitting model to feature number 57, HISDEG.PWY
#> 2023-09-09 09:30:50.254472 INFO::Fitting model to feature number 58, HISTSYN.PWY
#> 2023-09-09 09:30:50.258418 INFO::Fitting model to feature number 59, HOMOSER.METSYN.PWY
#> 2023-09-09 09:30:50.261436 INFO::Fitting model to feature number 60, HSERMETANA.PWY
#> 2023-09-09 09:30:50.263734 INFO::Fitting model to feature number 61, ILEUSYN.PWY
#> 2023-09-09 09:30:50.266044 INFO::Fitting model to feature number 62, KDO.NAGLIPASYN.PWY
#> 2023-09-09 09:30:50.268141 INFO::Fitting model to feature number 63, KETOGLUCONMET.PWY
#> 2023-09-09 09:30:50.270699 INFO::Fitting model to feature number 64, LACTOSECAT.PWY
#> 2023-09-09 09:30:50.274213 INFO::Fitting model to feature number 65, LEU.DEG2.PWY
#> 2023-09-09 09:30:50.2767 INFO::Fitting model to feature number 66, LIPASYN.PWY
#> 2023-09-09 09:30:50.278852 INFO::Fitting model to feature number 67, MET.SAM.PWY
#> 2023-09-09 09:30:50.281072 INFO::Fitting model to feature number 68, METH.ACETATE.PWY
#> 2023-09-09 09:30:50.283322 INFO::Fitting model to feature number 69, METHGLYUT.PWY
#> 2023-09-09 09:30:50.2855 INFO::Fitting model to feature number 70, NAD.BIOSYNTHESIS.II
#> 2023-09-09 09:30:50.287664 INFO::Fitting model to feature number 71, NAGLIPASYN.PWY
#> 2023-09-09 09:30:50.290516 INFO::Fitting model to feature number 72, NONMEVIPP.PWY
#> 2023-09-09 09:30:50.293539 INFO::Fitting model to feature number 73, NONOXIPENT.PWY
#> 2023-09-09 09:30:50.295919 INFO::Fitting model to feature number 74, OANTIGEN.PWY
#> 2023-09-09 09:30:50.298237 INFO::Fitting model to feature number 75, ORNARGDEG.PWY
#> 2023-09-09 09:30:50.300379 INFO::Fitting model to feature number 76, ORNDEG.PWY
#> 2023-09-09 09:30:50.302525 INFO::Fitting model to feature number 77, P105.PWY
#> 2023-09-09 09:30:50.30466 INFO::Fitting model to feature number 78, P108.PWY
#> 2023-09-09 09:30:50.307709 INFO::Fitting model to feature number 79, P122.PWY
#> 2023-09-09 09:30:50.310275 INFO::Fitting model to feature number 80, P124.PWY
#> 2023-09-09 09:30:50.312707 INFO::Fitting model to feature number 81, P125.PWY
#> 2023-09-09 09:30:50.315034 INFO::Fitting model to feature number 82, P161.PWY
#> 2023-09-09 09:30:50.317212 INFO::Fitting model to feature number 83, P162.PWY
#> 2023-09-09 09:30:50.319399 INFO::Fitting model to feature number 84, P163.PWY
#> 2023-09-09 09:30:50.321849 INFO::Fitting model to feature number 85, P164.PWY
#> 2023-09-09 09:30:50.325047 INFO::Fitting model to feature number 86, P23.PWY
#> 2023-09-09 09:30:50.327709 INFO::Fitting model to feature number 87, P341.PWY
#> 2023-09-09 09:30:50.332313 INFO::Fitting model to feature number 88, P4.PWY
#> 2023-09-09 09:30:50.337388 INFO::Fitting model to feature number 89, P42.PWY
#> 2023-09-09 09:30:50.343224 INFO::Fitting model to feature number 90, P441.PWY
#> 2023-09-09 09:30:50.347097 INFO::Fitting model to feature number 91, P461.PWY
#> 2023-09-09 09:30:50.349846 INFO::Fitting model to feature number 92, P562.PWY
#> 2023-09-09 09:30:50.353863 INFO::Fitting model to feature number 93, PANTO.PWY
#> 2023-09-09 09:30:50.358296 INFO::Fitting model to feature number 94, PANTOSYN.PWY
#> 2023-09-09 09:30:50.361492 INFO::Fitting model to feature number 95, PENTOSE.P.PWY
#> 2023-09-09 09:30:50.36436 INFO::Fitting model to feature number 96, PEPTIDOGLYCANSYN.PWY
#> 2023-09-09 09:30:50.367097 INFO::Fitting model to feature number 97, PHOSLIPSYN.PWY
#> 2023-09-09 09:30:50.3764 INFO::Fitting model to feature number 98, POLYAMINSYN3.PWY
#> 2023-09-09 09:30:50.383148 INFO::Fitting model to feature number 99, POLYAMSYN.PWY
#> 2023-09-09 09:30:50.386463 INFO::Fitting model to feature number 100, POLYISOPRENSYN.PWY
#> 2023-09-09 09:30:50.38949 INFO::Fitting model to feature number 101, PPGPPMET.PWY
#> 2023-09-09 09:30:50.393006 INFO::Fitting model to feature number 102, PRPP.PWY
#> 2023-09-09 09:30:50.395796 INFO::Fitting model to feature number 103, PWY.1269
#> 2023-09-09 09:30:50.398273 INFO::Fitting model to feature number 104, PWY.181
#> 2023-09-09 09:30:50.400436 INFO::Fitting model to feature number 105, PWY.1861
#> 2023-09-09 09:30:50.402649 INFO::Fitting model to feature number 106, PWY.2941
#> 2023-09-09 09:30:50.405417 INFO::Fitting model to feature number 107, PWY.2942
#> 2023-09-09 09:30:50.408732 INFO::Fitting model to feature number 108, PWY.3001
#> 2023-09-09 09:30:50.412628 INFO::Fitting model to feature number 109, PWY.3781
#> 2023-09-09 09:30:50.426864 INFO::Fitting model to feature number 110, PWY.4984
#> 2023-09-09 09:30:50.430492 INFO::Fitting model to feature number 111, PWY.5005
#> 2023-09-09 09:30:50.433407 INFO::Fitting model to feature number 112, PWY.5022
#> 2023-09-09 09:30:50.43617 INFO::Fitting model to feature number 113, PWY.5088
#> 2023-09-09 09:30:50.439497 INFO::Fitting model to feature number 114, PWY.5097
#> 2023-09-09 09:30:50.443115 INFO::Fitting model to feature number 115, PWY.5100
#> 2023-09-09 09:30:50.445934 INFO::Fitting model to feature number 116, PWY.5101
#> 2023-09-09 09:30:50.448214 INFO::Fitting model to feature number 117, PWY.5103
#> 2023-09-09 09:30:50.450387 INFO::Fitting model to feature number 118, PWY.5104
#> 2023-09-09 09:30:50.452895 INFO::Fitting model to feature number 119, PWY.5121
#> 2023-09-09 09:30:50.455329 INFO::Fitting model to feature number 120, PWY.5154
#> 2023-09-09 09:30:50.458985 INFO::Fitting model to feature number 121, PWY.5177
#> 2023-09-09 09:30:50.462157 INFO::Fitting model to feature number 122, PWY.5180
#> 2023-09-09 09:30:50.464472 INFO::Fitting model to feature number 123, PWY.5182
#> 2023-09-09 09:30:50.466736 INFO::Fitting model to feature number 124, PWY.5188
#> 2023-09-09 09:30:50.468912 INFO::Fitting model to feature number 125, PWY.5189
#> 2023-09-09 09:30:50.471495 INFO::Fitting model to feature number 126, PWY.5265
#> 2023-09-09 09:30:50.475051 INFO::Fitting model to feature number 127, PWY.5304
#> 2023-09-09 09:30:50.477663 INFO::Fitting model to feature number 128, PWY.5345
#> 2023-09-09 09:30:50.479858 INFO::Fitting model to feature number 129, PWY.5347
#> 2023-09-09 09:30:50.482073 INFO::Fitting model to feature number 130, PWY.5384
#> 2023-09-09 09:30:50.490608 INFO::Fitting model to feature number 131, PWY.5415
#> 2023-09-09 09:30:50.496693 INFO::Fitting model to feature number 132, PWY.5484
#> 2023-09-09 09:30:50.499767 INFO::Fitting model to feature number 133, PWY.5505
#> 2023-09-09 09:30:50.502432 INFO::Fitting model to feature number 134, PWY.5507
#> 2023-09-09 09:30:50.505713 INFO::Fitting model to feature number 135, PWY.5509
#> 2023-09-09 09:30:50.510153 INFO::Fitting model to feature number 136, PWY.5532
#> 2023-09-09 09:30:50.512646 INFO::Fitting model to feature number 137, PWY.5659
#> 2023-09-09 09:30:50.514871 INFO::Fitting model to feature number 138, PWY.5667
#> 2023-09-09 09:30:50.51711 INFO::Fitting model to feature number 139, PWY.5676
#> 2023-09-09 09:30:50.51934 INFO::Fitting model to feature number 140, PWY.5677
#> 2023-09-09 09:30:50.521736 INFO::Fitting model to feature number 141, PWY.5686
#> 2023-09-09 09:30:50.525071 INFO::Fitting model to feature number 142, PWY.5695
#> 2023-09-09 09:30:50.527553 INFO::Fitting model to feature number 143, PWY.5705
#> 2023-09-09 09:30:50.529676 INFO::Fitting model to feature number 144, PWY.5741
#> 2023-09-09 09:30:50.531771 INFO::Fitting model to feature number 145, PWY.5747
#> 2023-09-09 09:30:50.533891 INFO::Fitting model to feature number 146, PWY.5837
#> 2023-09-09 09:30:50.535985 INFO::Fitting model to feature number 147, PWY.5838
#> 2023-09-09 09:30:50.538527 INFO::Fitting model to feature number 148, PWY.5840
#> 2023-09-09 09:30:50.54183 INFO::Fitting model to feature number 149, PWY.5845
#> 2023-09-09 09:30:50.545229 INFO::Fitting model to feature number 150, PWY.5850
#> 2023-09-09 09:30:50.549409 INFO::Fitting model to feature number 151, PWY.5855
#> 2023-09-09 09:30:50.560364 INFO::Fitting model to feature number 152, PWY.5856
#> 2023-09-09 09:30:50.572037 INFO::Fitting model to feature number 153, PWY.5857
#> 2023-09-09 09:30:50.577477 INFO::Fitting model to feature number 154, PWY.5860
#> 2023-09-09 09:30:50.582796 INFO::Fitting model to feature number 155, PWY.5861
#> 2023-09-09 09:30:50.585985 INFO::Fitting model to feature number 156, PWY.5862
#> 2023-09-09 09:30:50.590027 INFO::Fitting model to feature number 157, PWY.5863
#> 2023-09-09 09:30:50.595146 INFO::Fitting model to feature number 158, PWY.5896
#> 2023-09-09 09:30:50.59887 INFO::Fitting model to feature number 159, PWY.5897
#> 2023-09-09 09:30:50.601722 INFO::Fitting model to feature number 160, PWY.5898
#> 2023-09-09 09:30:50.604803 INFO::Fitting model to feature number 161, PWY.5899
#> 2023-09-09 09:30:50.607959 INFO::Fitting model to feature number 162, PWY.5910
#> 2023-09-09 09:30:50.611333 INFO::Fitting model to feature number 163, PWY.5913
#> 2023-09-09 09:30:50.613925 INFO::Fitting model to feature number 164, PWY.5918
#> 2023-09-09 09:30:50.616172 INFO::Fitting model to feature number 165, PWY.5920
#> 2023-09-09 09:30:50.618368 INFO::Fitting model to feature number 166, PWY.5971
#> 2023-09-09 09:30:50.621981 INFO::Fitting model to feature number 167, PWY.5973
#> 2023-09-09 09:30:50.625035 INFO::Fitting model to feature number 168, PWY.5989
#> 2023-09-09 09:30:50.62734 INFO::Fitting model to feature number 169, PWY.6121
#> 2023-09-09 09:30:50.6294 INFO::Fitting model to feature number 170, PWY.6122
#> 2023-09-09 09:30:50.63168 INFO::Fitting model to feature number 171, PWY.6123
#> 2023-09-09 09:30:50.633893 INFO::Fitting model to feature number 172, PWY.6125
#> 2023-09-09 09:30:50.636115 INFO::Fitting model to feature number 173, PWY.6126
#> 2023-09-09 09:30:50.638291 INFO::Fitting model to feature number 174, PWY.6147
#> 2023-09-09 09:30:50.641356 INFO::Fitting model to feature number 175, PWY.6151
#> 2023-09-09 09:30:50.644431 INFO::Fitting model to feature number 176, PWY.6163
#> 2023-09-09 09:30:50.646893 INFO::Fitting model to feature number 177, PWY.621
#> 2023-09-09 09:30:50.649061 INFO::Fitting model to feature number 178, PWY.6263
#> 2023-09-09 09:30:50.651268 INFO::Fitting model to feature number 179, PWY.6269
#> 2023-09-09 09:30:50.653904 INFO::Fitting model to feature number 180, PWY.6277
#> 2023-09-09 09:30:50.657236 INFO::Fitting model to feature number 181, PWY.6282
#> 2023-09-09 09:30:50.659776 INFO::Fitting model to feature number 182, PWY.6317
#> 2023-09-09 09:30:50.662209 INFO::Fitting model to feature number 183, PWY.6353
#> 2023-09-09 09:30:50.664407 INFO::Fitting model to feature number 184, PWY.6385
#> 2023-09-09 09:30:50.666535 INFO::Fitting model to feature number 185, PWY.6386
#> 2023-09-09 09:30:50.66876 INFO::Fitting model to feature number 186, PWY.6387
#> 2023-09-09 09:30:50.670944 INFO::Fitting model to feature number 187, PWY.6396
#> 2023-09-09 09:30:50.675598 INFO::Fitting model to feature number 188, PWY.6467
#> 2023-09-09 09:30:50.679087 INFO::Fitting model to feature number 189, PWY.6470
#> 2023-09-09 09:30:50.681534 INFO::Fitting model to feature number 190, PWY.6471
#> 2023-09-09 09:30:50.683778 INFO::Fitting model to feature number 191, PWY.6478
#> 2023-09-09 09:30:50.686166 INFO::Fitting model to feature number 192, PWY.6507
#> 2023-09-09 09:30:50.688589 INFO::Fitting model to feature number 193, PWY.6519
#> 2023-09-09 09:30:50.691828 INFO::Fitting model to feature number 194, PWY.6545
#> 2023-09-09 09:30:50.694749 INFO::Fitting model to feature number 195, PWY.6572
#> 2023-09-09 09:30:50.697543 INFO::Fitting model to feature number 196, PWY.6588
#> 2023-09-09 09:30:50.700112 INFO::Fitting model to feature number 197, PWY.6590
#> 2023-09-09 09:30:50.702315 INFO::Fitting model to feature number 198, PWY.6608
#> 2023-09-09 09:30:50.706472 INFO::Fitting model to feature number 199, PWY.6609
#> 2023-09-09 09:30:50.711385 INFO::Fitting model to feature number 200, PWY.6612
#> 2023-09-09 09:30:50.713933 INFO::Fitting model to feature number 201, PWY.6628
#> 2023-09-09 09:30:50.716401 INFO::Fitting model to feature number 202, PWY.6629
#> 2023-09-09 09:30:50.71925 INFO::Fitting model to feature number 203, PWY.6630
#> 2023-09-09 09:30:50.725964 INFO::Fitting model to feature number 204, PWY.6641
#> 2023-09-09 09:30:50.729033 INFO::Fitting model to feature number 205, PWY.6690
#> 2023-09-09 09:30:50.731442 INFO::Fitting model to feature number 206, PWY.6700
#> 2023-09-09 09:30:50.73384 INFO::Fitting model to feature number 207, PWY.6703
#> 2023-09-09 09:30:50.73603 INFO::Fitting model to feature number 208, PWY.6708
#> 2023-09-09 09:30:50.739045 INFO::Fitting model to feature number 209, PWY.6728
#> 2023-09-09 09:30:50.742576 INFO::Fitting model to feature number 210, PWY.6737
#> 2023-09-09 09:30:50.745224 INFO::Fitting model to feature number 211, PWY.6749
#> 2023-09-09 09:30:50.7474 INFO::Fitting model to feature number 212, PWY.6891
#> 2023-09-09 09:30:50.749592 INFO::Fitting model to feature number 213, PWY.6892
#> 2023-09-09 09:30:50.751753 INFO::Fitting model to feature number 214, PWY.6895
#> 2023-09-09 09:30:50.753902 INFO::Fitting model to feature number 215, PWY.6897
#> 2023-09-09 09:30:50.757502 INFO::Fitting model to feature number 216, PWY.6901
#> 2023-09-09 09:30:50.760754 INFO::Fitting model to feature number 217, PWY.6969
#> 2023-09-09 09:30:50.763482 INFO::Fitting model to feature number 218, PWY.7003
#> 2023-09-09 09:30:50.765812 INFO::Fitting model to feature number 219, PWY.7013
#> 2023-09-09 09:30:50.768021 INFO::Fitting model to feature number 220, PWY.7090
#> 2023-09-09 09:30:50.770468 INFO::Fitting model to feature number 221, PWY.7094
#> 2023-09-09 09:30:50.775148 INFO::Fitting model to feature number 222, PWY.7111
#> 2023-09-09 09:30:50.778034 INFO::Fitting model to feature number 223, PWY.7184
#> 2023-09-09 09:30:50.780378 INFO::Fitting model to feature number 224, PWY.7187
#> 2023-09-09 09:30:50.782734 INFO::Fitting model to feature number 225, PWY.7196
#> 2023-09-09 09:30:50.784859 INFO::Fitting model to feature number 226, PWY.7197
#> 2023-09-09 09:30:50.78742 INFO::Fitting model to feature number 227, PWY.7199
#> 2023-09-09 09:30:50.791869 INFO::Fitting model to feature number 228, PWY.7200
#> 2023-09-09 09:30:50.794684 INFO::Fitting model to feature number 229, PWY.7208
#> 2023-09-09 09:30:50.799217 INFO::Fitting model to feature number 230, PWY.7211
#> 2023-09-09 09:30:50.803401 INFO::Fitting model to feature number 231, PWY.7219
#> 2023-09-09 09:30:50.809876 INFO::Fitting model to feature number 232, PWY.7220
#> 2023-09-09 09:30:50.813929 INFO::Fitting model to feature number 233, PWY.7221
#> 2023-09-09 09:30:50.817831 INFO::Fitting model to feature number 234, PWY.7222
#> 2023-09-09 09:30:50.821386 INFO::Fitting model to feature number 235, PWY.7228
#> 2023-09-09 09:30:50.826672 INFO::Fitting model to feature number 236, PWY.7229
#> 2023-09-09 09:30:50.829956 INFO::Fitting model to feature number 237, PWY.7234
#> 2023-09-09 09:30:50.832664 INFO::Fitting model to feature number 238, PWY.7237
#> 2023-09-09 09:30:50.835115 INFO::Fitting model to feature number 239, PWY.7242
#> 2023-09-09 09:30:50.839223 INFO::Fitting model to feature number 240, PWY.7254
#> 2023-09-09 09:30:50.841996 INFO::Fitting model to feature number 241, PWY.7315
#> 2023-09-09 09:30:50.844635 INFO::Fitting model to feature number 242, PWY.7323
#> 2023-09-09 09:30:50.846884 INFO::Fitting model to feature number 243, PWY.7328
#> 2023-09-09 09:30:50.848986 INFO::Fitting model to feature number 244, PWY.7332
#> 2023-09-09 09:30:50.851059 INFO::Fitting model to feature number 245, PWY.7347
#> 2023-09-09 09:30:50.853654 INFO::Fitting model to feature number 246, PWY.7371
#> 2023-09-09 09:30:50.85668 INFO::Fitting model to feature number 247, PWY.7373
#> 2023-09-09 09:30:50.859064 INFO::Fitting model to feature number 248, PWY.7374
#> 2023-09-09 09:30:50.861154 INFO::Fitting model to feature number 249, PWY.7377
#> 2023-09-09 09:30:50.863176 INFO::Fitting model to feature number 250, PWY.7392
#> 2023-09-09 09:30:50.865221 INFO::Fitting model to feature number 251, PWY.7400
#> 2023-09-09 09:30:50.867277 INFO::Fitting model to feature number 252, PWY.7431
#> 2023-09-09 09:30:50.869325 INFO::Fitting model to feature number 253, PWY.7446
#> 2023-09-09 09:30:50.87141 INFO::Fitting model to feature number 254, PWY.7456
#> 2023-09-09 09:30:50.874123 INFO::Fitting model to feature number 255, PWY.7539
#> 2023-09-09 09:30:50.877196 INFO::Fitting model to feature number 256, PWY.7560
#> 2023-09-09 09:30:50.879743 INFO::Fitting model to feature number 257, PWY.7663
#> 2023-09-09 09:30:50.882145 INFO::Fitting model to feature number 258, PWY.7664
#> 2023-09-09 09:30:50.884263 INFO::Fitting model to feature number 259, PWY.841
#> 2023-09-09 09:30:50.886319 INFO::Fitting model to feature number 260, PWY.922
#> 2023-09-09 09:30:50.890116 INFO::Fitting model to feature number 261, PWY0.1061
#> 2023-09-09 09:30:50.893057 INFO::Fitting model to feature number 262, PWY0.1241
#> 2023-09-09 09:30:50.895307 INFO::Fitting model to feature number 263, PWY0.1261
#> 2023-09-09 09:30:50.897422 INFO::Fitting model to feature number 264, PWY0.1277
#> 2023-09-09 09:30:50.899529 INFO::Fitting model to feature number 265, PWY0.1296
#> 2023-09-09 09:30:50.901626 INFO::Fitting model to feature number 266, PWY0.1297
#> 2023-09-09 09:30:50.903744 INFO::Fitting model to feature number 267, PWY0.1298
#> 2023-09-09 09:30:50.906092 INFO::Fitting model to feature number 268, PWY0.1319
#> 2023-09-09 09:30:50.908569 INFO::Fitting model to feature number 269, PWY0.1338
#> 2023-09-09 09:30:50.910787 INFO::Fitting model to feature number 270, PWY0.1415
#> 2023-09-09 09:30:50.912897 INFO::Fitting model to feature number 271, PWY0.1479
#> 2023-09-09 09:30:50.915039 INFO::Fitting model to feature number 272, PWY0.1533
#> 2023-09-09 09:30:50.917145 INFO::Fitting model to feature number 273, PWY0.1586
#> 2023-09-09 09:30:50.919241 INFO::Fitting model to feature number 274, PWY0.162
#> 2023-09-09 09:30:50.923128 INFO::Fitting model to feature number 275, PWY0.166
#> 2023-09-09 09:30:50.92884 INFO::Fitting model to feature number 276, PWY0.41
#> 2023-09-09 09:30:50.932624 INFO::Fitting model to feature number 277, PWY0.42
#> 2023-09-09 09:30:50.936876 INFO::Fitting model to feature number 278, PWY0.781
#> 2023-09-09 09:30:50.939815 INFO::Fitting model to feature number 279, PWY0.845
#> 2023-09-09 09:30:50.942231 INFO::Fitting model to feature number 280, PWY0.862
#> 2023-09-09 09:30:50.944388 INFO::Fitting model to feature number 281, PWY490.3
#> 2023-09-09 09:30:50.946392 INFO::Fitting model to feature number 282, PWY4FS.7
#> 2023-09-09 09:30:50.948408 INFO::Fitting model to feature number 283, PWY4FS.8
#> 2023-09-09 09:30:50.950499 INFO::Fitting model to feature number 284, PWYG.321
#> 2023-09-09 09:30:50.952698 INFO::Fitting model to feature number 285, PYRIDNUCSAL.PWY
#> 2023-09-09 09:30:50.95493 INFO::Fitting model to feature number 286, PYRIDNUCSYN.PWY
#> 2023-09-09 09:30:50.957276 INFO::Fitting model to feature number 287, PYRIDOXSYN.PWY
#> 2023-09-09 09:30:50.959513 INFO::Fitting model to feature number 288, REDCITCYC
#> 2023-09-09 09:30:50.961579 INFO::Fitting model to feature number 289, RHAMCAT.PWY
#> 2023-09-09 09:30:50.963623 INFO::Fitting model to feature number 290, RIBOSYN2.PWY
#> 2023-09-09 09:30:50.965713 INFO::Fitting model to feature number 291, RUMP.PWY
#> 2023-09-09 09:30:51.158138 INFO::Fitting model to feature number 292, SALVADEHYPOX.PWY
#> 2023-09-09 09:30:51.162292 INFO::Fitting model to feature number 293, SER.GLYSYN.PWY
#> 2023-09-09 09:30:51.164642 INFO::Fitting model to feature number 294, SO4ASSIM.PWY
#> 2023-09-09 09:30:51.1668 INFO::Fitting model to feature number 295, SUCSYN.PWY
#> 2023-09-09 09:30:51.168989 INFO::Fitting model to feature number 296, SULFATE.CYS.PWY
#> 2023-09-09 09:30:51.171318 INFO::Fitting model to feature number 297, TCA
#> 2023-09-09 09:30:51.173916 INFO::Fitting model to feature number 298, TCA.GLYOX.BYPASS
#> 2023-09-09 09:30:51.176633 INFO::Fitting model to feature number 299, TEICHOICACID.PWY
#> 2023-09-09 09:30:51.179362 INFO::Fitting model to feature number 300, THISYN.PWY
#> 2023-09-09 09:30:51.181949 INFO::Fitting model to feature number 301, THRESYN.PWY
#> 2023-09-09 09:30:51.18442 INFO::Fitting model to feature number 302, TRNA.CHARGING.PWY
#> 2023-09-09 09:30:51.186828 INFO::Fitting model to feature number 303, TRPSYN.PWY
#> 2023-09-09 09:30:51.18931 INFO::Fitting model to feature number 304, TYRFUMCAT.PWY
#> 2023-09-09 09:30:51.191939 INFO::Fitting model to feature number 305, UBISYN.PWY
#> 2023-09-09 09:30:51.194416 INFO::Fitting model to feature number 306, UDPNAGSYN.PWY
#> 2023-09-09 09:30:51.196699 INFO::Fitting model to feature number 307, VALDEG.PWY
#> 2023-09-09 09:30:51.199036 INFO::Fitting model to feature number 308, VALSYN.PWY
#> 2023-09-09 09:30:51.245527 INFO::Counting total values for each feature
#> 2023-09-09 09:30:51.260589 INFO::Writing filtered data to file Maaslin2_results_Environment/features/filtered_data.tsv
#> 2023-09-09 09:30:51.295497 INFO::Writing filtered, normalized data to file Maaslin2_results_Environment/features/filtered_data_norm.tsv
#> 2023-09-09 09:30:51.325636 INFO::Writing filtered, normalized, transformed data to file Maaslin2_results_Environment/features/filtered_data_norm_transformed.tsv
#> 2023-09-09 09:30:51.356622 WARNING::Deleting existing residuals file: Maaslin2_results_Environment/fits/residuals.rds
#> 2023-09-09 09:30:51.359195 INFO::Writing residuals to file Maaslin2_results_Environment/fits/residuals.rds
#> 2023-09-09 09:30:51.365479 WARNING::Deleting existing fitted file: Maaslin2_results_Environment/fits/fitted.rds
#> 2023-09-09 09:30:51.366679 INFO::Writing fitted values to file Maaslin2_results_Environment/fits/fitted.rds
#> 2023-09-09 09:30:51.368409 INFO::Writing all results to file (ordered by increasing q-values): Maaslin2_results_Environment/all_results.tsv
#> 2023-09-09 09:30:51.374954 INFO::Writing the significant results (those which are less than or equal to the threshold of 0.250000 ) to file (ordered by increasing q-values): Maaslin2_results_Environment/significant_results.tsv
#> 2023-09-09 09:30:51.378158 INFO::Writing heatmap of significant results to file: Maaslin2_results_Environment/heatmap.pdf
#> [1] "There is not enough metadata in the associations to create a heatmap plot. Please review the associations in text output file."
#> 2023-09-09 09:30:51.389407 INFO::Writing association plots (one for each significant association) to output folder: Maaslin2_results_Environment
#> 2023-09-09 09:30:51.399763 INFO::Plotting associations from most to least significant, grouped by metadata
#> 2023-09-09 09:30:51.401023 INFO::Plotting data for metadata number 1, Environment
#> 2023-09-09 09:30:51.448327 INFO::Creating boxplot for categorical data, Environment vs PWY0.1241
#> 2023-09-09 09:30:51.750926 INFO::Creating boxplot for categorical data, Environment vs PWY.4984
#> 2023-09-09 09:30:51.907126 INFO::Creating boxplot for categorical data, Environment vs PWY.7373
#> 2023-09-09 09:30:52.220776 INFO::Creating boxplot for categorical data, Environment vs PWY0.781
#> 2023-09-09 09:30:52.407589 INFO::Creating boxplot for categorical data, Environment vs PPGPPMET.PWY
#> 2023-09-09 09:30:53.407373 INFO::Creating boxplot for categorical data, Environment vs PWY.3781
#> 2023-09-09 09:30:53.714003 INFO::Creating boxplot for categorical data, Environment vs PWY.5747
#> 2023-09-09 09:30:53.870944 INFO::Creating boxplot for categorical data, Environment vs PWY0.42
#> 2023-09-09 09:30:54.060496 INFO::Creating boxplot for categorical data, Environment vs PWY0.1586
#> 2023-09-09 09:30:54.192217 INFO::Creating boxplot for categorical data, Environment vs HEMESYN2.PWY
#> 2023-09-09 09:30:55.652733 INFO::Creating boxplot for categorical data, Environment vs P4.PWY
#> 2023-09-09 09:30:55.80308 INFO::Creating boxplot for categorical data, Environment vs PHOSLIPSYN.PWY
#> 2023-09-09 09:30:56.11065 INFO::Creating boxplot for categorical data, Environment vs PWY.6317
#> 2023-09-09 09:30:56.6197 INFO::Creating boxplot for categorical data, Environment vs PWY.7254
#> 2023-09-09 09:30:56.836362 INFO::Creating boxplot for categorical data, Environment vs REDCITCYC
#> 2023-09-09 09:30:57.150714 INFO::Creating boxplot for categorical data, Environment vs PWY.6641
#> 2023-09-09 09:30:57.506632 INFO::Creating boxplot for categorical data, Environment vs PWY.6263
#> 2023-09-09 09:30:57.702283 INFO::Creating boxplot for categorical data, Environment vs PWY.7371
#> 2023-09-09 09:30:57.881074 INFO::Creating boxplot for categorical data, Environment vs MET.SAM.PWY
#> 2023-09-09 09:30:58.098214 INFO::Creating boxplot for categorical data, Environment vs ASPASN.PWY
#> 2023-09-09 09:30:58.292819 INFO::Creating boxplot for categorical data, Environment vs HOMOSER.METSYN.PWY
#> 2023-09-09 09:30:58.516674 INFO::Creating boxplot for categorical data, Environment vs HSERMETANA.PWY
#> 2023-09-09 09:30:58.790571 INFO::Creating boxplot for categorical data, Environment vs PWY.5347
#> 2023-09-09 09:30:58.930214 INFO::Creating boxplot for categorical data, Environment vs PWY.5667
#> 2023-09-09 09:30:59.146431 INFO::Creating boxplot for categorical data, Environment vs PWY0.1319
#> 2023-09-09 09:30:59.375031 INFO::Creating boxplot for categorical data, Environment vs ANAGLYCOLYSIS.PWY
#> 2023-09-09 09:30:59.570951 INFO::Creating boxplot for categorical data, Environment vs PWY.5686
#> 2023-09-09 09:30:59.892037 INFO::Creating boxplot for categorical data, Environment vs PWY0.1298
#> 2023-09-09 09:31:00.149536 INFO::Creating boxplot for categorical data, Environment vs PWY.621
#> 2023-09-09 09:31:00.302818 INFO::Creating boxplot for categorical data, Environment vs P23.PWY
#> 2023-09-09 09:31:00.447629 INFO::Creating boxplot for categorical data, Environment vs PWY0.1297
#> 2023-09-09 09:31:00.5723 INFO::Creating boxplot for categorical data, Environment vs PWY.5855
#> 2023-09-09 09:31:00.73074 INFO::Creating boxplot for categorical data, Environment vs PWY.5856
#> 2023-09-09 09:31:00.963094 INFO::Creating boxplot for categorical data, Environment vs PWY.5857
#> 2023-09-09 09:31:01.069568 INFO::Creating boxplot for categorical data, Environment vs PWY.6708
#> 2023-09-09 09:31:01.23326 INFO::Creating boxplot for categorical data, Environment vs PWY4FS.7
#> 2023-09-09 09:31:01.379024 INFO::Creating boxplot for categorical data, Environment vs PWY4FS.8
#> 2023-09-09 09:31:01.513603 INFO::Creating boxplot for categorical data, Environment vs UBISYN.PWY
#> 2023-09-09 09:31:01.786527 INFO::Creating boxplot for categorical data, Environment vs PWY.6609
#> 2023-09-09 09:31:01.956875 INFO::Creating boxplot for categorical data, Environment vs GLUCOSE1PMETAB.PWY
#> 2023-09-09 09:31:02.28785 INFO::Creating boxplot for categorical data, Environment vs GLYCOGENSYNTH.PWY
#> 2023-09-09 09:31:02.56037 INFO::Creating boxplot for categorical data, Environment vs NONOXIPENT.PWY
#> 2023-09-09 09:31:02.712739 INFO::Creating boxplot for categorical data, Environment vs P562.PWY
#> 2023-09-09 09:31:02.870645 INFO::Creating boxplot for categorical data, Environment vs PWY.5100
#> 2023-09-09 09:31:03.274675 INFO::Creating boxplot for categorical data, Environment vs PWY.5837
#> 2023-09-09 09:31:03.428603 INFO::Creating boxplot for categorical data, Environment vs PWY.5838
#> 2023-09-09 09:31:04.000515 INFO::Creating boxplot for categorical data, Environment vs PWY.5840
#> 2023-09-09 09:31:04.229888 INFO::Creating boxplot for categorical data, Environment vs PWY.5861
#> 2023-09-09 09:31:04.387265 INFO::Creating boxplot for categorical data, Environment vs PWY.5863
#> 2023-09-09 09:31:04.633905 INFO::Creating boxplot for categorical data, Environment vs PWY.5897
#> 2023-09-09 09:31:04.768958 INFO::Creating boxplot for categorical data, Environment vs PWY.5898
#> 2023-09-09 09:31:04.928878 INFO::Creating boxplot for categorical data, Environment vs PWY.5899
#> 2023-09-09 09:31:05.06424 INFO::Creating boxplot for categorical data, Environment vs PWY.7208
#> 2023-09-09 09:31:05.182089 INFO::Creating boxplot for categorical data, Environment vs PWY.7219
#> 2023-09-09 09:31:05.314666 INFO::Creating boxplot for categorical data, Environment vs PWY.7323
#> 2023-09-09 09:31:05.42171 INFO::Creating boxplot for categorical data, Environment vs PWY0.1296
#> 2023-09-09 09:31:05.524776 INFO::Creating boxplot for categorical data, Environment vs DTDPRHAMSYN.PWY
#> 2023-09-09 09:31:05.636548 INFO::Creating boxplot for categorical data, Environment vs CENTFERM.PWY
#> 2023-09-09 09:31:05.753118 INFO::Creating boxplot for categorical data, Environment vs GLYCOLYSIS.E.D
#> 2023-09-09 09:31:05.868258 INFO::Creating boxplot for categorical data, Environment vs P161.PWY
#> 2023-09-09 09:31:06.003201 INFO::Creating boxplot for categorical data, Environment vs POLYAMINSYN3.PWY
#> 2023-09-09 09:31:06.104046 INFO::Creating boxplot for categorical data, Environment vs PWY.1269
#> 2023-09-09 09:31:06.208309 INFO::Creating boxplot for categorical data, Environment vs PWY.6590
#> 2023-09-09 09:31:06.307048 INFO::Creating boxplot for categorical data, Environment vs PWY.5918
#> 2023-09-09 09:31:06.406412 INFO::Creating boxplot for categorical data, Environment vs TRNA.CHARGING.PWY
#> 2023-09-09 09:31:06.529907 INFO::Creating boxplot for categorical data, Environment vs COLANSYN.PWY
#> 2023-09-09 09:31:06.650477 INFO::Creating boxplot for categorical data, Environment vs PWY.6467
#> 2023-09-09 09:31:06.749444 INFO::Creating boxplot for categorical data, Environment vs NAGLIPASYN.PWY
#> 2023-09-09 09:31:06.850312 INFO::Creating boxplot for categorical data, Environment vs PWY.7374
#> 2023-09-09 09:31:06.993649 INFO::Creating boxplot for categorical data, Environment vs PWY.181
#> 2023-09-09 09:31:07.113043 INFO::Creating boxplot for categorical data, Environment vs FAO.PWY
#> 2023-09-09 09:31:07.306145 INFO::Creating boxplot for categorical data, Environment vs THISYN.PWY
#> 2023-09-09 09:31:07.402428 INFO::Creating boxplot for categorical data, Environment vs HEME.BIOSYNTHESIS.II
#> 2023-09-09 09:31:07.510858 INFO::Creating boxplot for categorical data, Environment vs FUCCAT.PWY
#> 2023-09-09 09:31:07.625488 INFO::Creating boxplot for categorical data, Environment vs GLYCOLYSIS.TCA.GLYOX.BYPASS
#> 2023-09-09 09:31:07.743097 INFO::Creating boxplot for categorical data, Environment vs PWY.7315
#> 2023-09-09 09:31:07.994907 INFO::Creating boxplot for categorical data, Environment vs PWY.5920
#> 2023-09-09 09:31:08.158694 INFO::Creating boxplot for categorical data, Environment vs GALACTUROCAT.PWY
#> 2023-09-09 09:31:08.298752 INFO::Creating boxplot for categorical data, Environment vs TCA.GLYOX.BYPASS
#> 2023-09-09 09:31:08.466503 INFO::Creating boxplot for categorical data, Environment vs GLCMANNANAUT.PWY
#> 2023-09-09 09:31:08.602243 INFO::Creating boxplot for categorical data, Environment vs PWY.5177
#> 2023-09-09 09:31:08.740593 INFO::Creating boxplot for categorical data, Environment vs PWY.5505
#> 2023-09-09 09:31:08.869889 INFO::Creating boxplot for categorical data, Environment vs PWY.5695
#> 2023-09-09 09:31:08.985123 INFO::Creating boxplot for categorical data, Environment vs PWY0.41
#> 2023-09-09 09:31:09.08398 INFO::Creating boxplot for categorical data, Environment vs P441.PWY
#> 2023-09-09 09:31:09.212441 INFO::Creating boxplot for categorical data, Environment vs GALACT.GLUCUROCAT.PWY
#> 2023-09-09 09:31:09.360832 INFO::Creating boxplot for categorical data, Environment vs PWY.7197
#> 2023-09-09 09:31:09.622898 INFO::Creating boxplot for categorical data, Environment vs PWY0.1479
#> 2023-09-09 09:31:09.840985 INFO::Creating boxplot for categorical data, Environment vs BIOTIN.BIOSYNTHESIS.PWY
#> 2023-09-09 09:31:10.081385 INFO::Creating boxplot for categorical data, Environment vs PWY0.1415
#> 2023-09-09 09:31:10.230797 INFO::Creating boxplot for categorical data, Environment vs PWY.6519
#> 2023-09-09 09:31:10.405661 INFO::Creating boxplot for categorical data, Environment vs COA.PWY
#> 2023-09-09 09:31:10.658935 INFO::Creating boxplot for categorical data, Environment vs DAPLYSINESYN.PWY
#> 2023-09-09 09:31:10.793034 INFO::Creating boxplot for categorical data, Environment vs FASYN.INITIAL.PWY
#> 2023-09-09 09:31:10.977526 INFO::Creating boxplot for categorical data, Environment vs FOLSYN.PWY
#> 2023-09-09 09:31:11.155835 INFO::Creating boxplot for categorical data, Environment vs FUC.RHAMCAT.PWY
#> 2023-09-09 09:31:11.321506 INFO::Creating boxplot for categorical data, Environment vs GLUCUROCAT.PWY
#> 2023-09-09 09:31:11.604938 INFO::Creating boxplot for categorical data, Environment vs PWY.5989
#> 2023-09-09 09:31:11.729347 INFO::Creating boxplot for categorical data, Environment vs PWY.6282
#> 2023-09-09 09:31:11.926072 INFO::Creating boxplot for categorical data, Environment vs PWY.6612
#> 2023-09-09 09:31:12.110795 INFO::Creating boxplot for categorical data, Environment vs PWY.7184
#> 2023-09-09 09:31:12.332899 INFO::Creating boxplot for categorical data, Environment vs PWY.7221
#> 2023-09-09 09:31:12.627976 INFO::Creating boxplot for categorical data, Environment vs PWY.7228
#> 2023-09-09 09:31:12.78076 INFO::Creating boxplot for categorical data, Environment vs PWY.7242
#> 2023-09-09 09:31:12.944412 INFO::Creating boxplot for categorical data, Environment vs PWY.7328
#> 2023-09-09 09:31:13.099222 INFO::Creating boxplot for categorical data, Environment vs PWY.7664
#> 2023-09-09 09:31:13.266866 INFO::Creating boxplot for categorical data, Environment vs PWY.841
#> 2023-09-09 09:31:13.435488 INFO::Creating boxplot for categorical data, Environment vs PWY0.862
#> 2023-09-09 09:31:13.5837 INFO::Creating boxplot for categorical data, Environment vs PWYG.321
#> 2023-09-09 09:31:13.682831 INFO::Creating boxplot for categorical data, Environment vs SER.GLYSYN.PWY
#> 2023-09-09 09:31:13.780938 INFO::Creating boxplot for categorical data, Environment vs PWY.6386
#> 2023-09-09 09:31:13.914337 INFO::Creating boxplot for categorical data, Environment vs PWY.5971
#> 2023-09-09 09:31:14.027425 INFO::Creating boxplot for categorical data, Environment vs PEPTIDOGLYCANSYN.PWY
#> 2023-09-09 09:31:14.156181 INFO::Creating boxplot for categorical data, Environment vs PWY.5659
#> 2023-09-09 09:31:14.255616 INFO::Creating boxplot for categorical data, Environment vs PWY.6125
#> 2023-09-09 09:31:14.406684 INFO::Creating boxplot for categorical data, Environment vs PWY.7196
#> 2023-09-09 09:31:14.534498 INFO::Creating boxplot for categorical data, Environment vs PWY0.162
#> 2023-09-09 09:31:14.7111 INFO::Creating boxplot for categorical data, Environment vs PWY.7211
#> 2023-09-09 09:31:14.938471 INFO::Creating boxplot for categorical data, Environment vs OANTIGEN.PWY
#> 2023-09-09 09:31:15.082742 INFO::Creating boxplot for categorical data, Environment vs PWY.6385
#> 2023-09-09 09:31:15.212178 INFO::Creating boxplot for categorical data, Environment vs PWY.6545
#> 2023-09-09 09:31:15.352272 INFO::Creating boxplot for categorical data, Environment vs CALVIN.PWY
#> 2023-09-09 09:31:15.460314 INFO::Creating boxplot for categorical data, Environment vs NAD.BIOSYNTHESIS.II
#> 2023-09-09 09:31:15.612808 INFO::Creating boxplot for categorical data, Environment vs PWY.6387
#> 2023-09-09 09:31:15.713187 INFO::Creating boxplot for categorical data, Environment vs GLUCONEO.PWY
#> 2023-09-09 09:31:15.808441 INFO::Creating boxplot for categorical data, Environment vs LIPASYN.PWY
#> 2023-09-09 09:31:15.909962 INFO::Creating boxplot for categorical data, Environment vs DENOVOPURINE2.PWY
#> 2023-09-09 09:31:16.007812 INFO::Creating boxplot for categorical data, Environment vs PWY.7663
#> 2023-09-09 09:31:16.1067 INFO::Creating boxplot for categorical data, Environment vs FASYN.ELONG.PWY
#> 2023-09-09 09:31:16.227949 INFO::Creating boxplot for categorical data, Environment vs HISDEG.PWY
#> 2023-09-09 09:31:16.32834 INFO::Creating boxplot for categorical data, Environment vs X1CMET2.PWY
#> 2023-09-09 09:31:16.549781 INFO::Creating boxplot for categorical data, Environment vs PWY0.1061
#> 2023-09-09 09:31:16.653725 INFO::Creating boxplot for categorical data, Environment vs GLYCOCAT.PWY
#> 2023-09-09 09:31:16.759119 INFO::Creating boxplot for categorical data, Environment vs PWY.7400
#> 2023-09-09 09:31:17.028603 INFO::Creating boxplot for categorical data, Environment vs ARGSYN.PWY
#> 2023-09-09 09:31:17.155677 INFO::Creating boxplot for categorical data, Environment vs PWY.5384
#> 2023-09-09 09:31:17.290539 INFO::Creating boxplot for categorical data, Environment vs PWY.7111
#> Maaslin2 analysis complete. You can view the full analysis results and logs in the current default file location: /Users/apple/Microbiome/ggpicrust2总/ggpicrust2/docs/reference/Maaslin2_results_Environment

names(daa_results_list) <- methods
# Correct Maaslin2 feature names by replacing dots with hyphens.
# Note: When using Maaslin2 as the differential abundance analysis method,
# it modifies the original feature names by replacing hyphens (-) with dots (.).
# This replacement can cause inconsistencies when trying to compare results from Maaslin2
# with those from other methods that do not modify feature names.
# Therefore, this line of code reverses that replacement, converting the dots back into
# hyphens for accurate and consistent comparisons across different methods.
daa_results_list[["Maaslin2"]]$feature <- gsub("\\.", "-", daa_results_list[["Maaslin2"]]$feature)

# Compare results across different methods
comparison_results <- compare_daa_results(daa_results_list = daa_results_list,
method_names = c("ALDEx2","DESeq2", "edgeR", "Maaslin2"))
#> Comparing 4 methods:
#> 
#> The ALDEx2 method obtained 122 statistically significant features.
#> The number of features that are common to other methods is 8 
#> The number of features that are different from other methods is 122 
#> The names of the features that are different from other methods are 1CMET2-PWY, ANAEROFRUCAT-PWY, ANAGLYCOLYSIS-PWY, ASPASN-PWY, CALVIN-PWY, COA-PWY, DAPLYSINESYN-PWY, DTDPRHAMSYN-PWY, GALACT-GLUCUROCAT-PWY, GALACTUROCAT-PWY, GLCMANNANAUT-PWY, GLUCONEO-PWY, GLUCOSE1PMETAB-PWY, GLUCUROCAT-PWY, GLYCOGENSYNTH-PWY, GLYCOLYSIS, NONOXIPENT-PWY, OANTIGEN-PWY, P161-PWY, PENTOSE-P-PWY, PEPTIDOGLYCANSYN-PWY, PHOSLIPSYN-PWY, POLYAMINSYN3-PWY, PPGPPMET-PWY, PWY-3781, PWY-4984, PWY-5100, PWY-5384, PWY-5484, PWY-5505, PWY-5667, PWY-5686, PWY-5747, PWY-5837, PWY-5838, PWY-5840, PWY-5845, PWY-5850, PWY-5860, PWY-5861, PWY-5862, PWY-5863, PWY-5896, PWY-5897, PWY-5898, PWY-5899, PWY-6151, PWY-621, PWY-6263, PWY-6317, PWY-6385, PWY-6386, PWY-6387, PWY-6471, PWY-6609, PWY-7199, PWY-7208, PWY-7219, PWY-7221, PWY-7229, PWY-7242, PWY-7315, PWY-7371, PWY-7373, PWY0-1241, PWY0-1296, PWY0-1297, PWY0-1298, PWY0-1319, PWY0-1586, PWY0-42, PWY4FS-7, PWY4FS-8, REDCITCYC, SER-GLYSYN-PWY, TRNA-CHARGING-PWY, UDPNAGSYN-PWY, ANAEROFRUCAT-PWY, GALACT-GLUCUROCAT-PWY, GALACTUROCAT-PWY, GLCMANNANAUT-PWY, GLUCOSE1PMETAB-PWY, GLUCUROCAT-PWY, GLYCOGENSYNTH-PWY, P161-PWY, PENTOSE-P-PWY, POLYAMINSYN3-PWY, PPGPPMET-PWY, PWY-3781, PWY-4984, PWY-5100, PWY-5384, PWY-5484, PWY-5505, PWY-5747, PWY-5837, PWY-5838, PWY-5840, PWY-5861, PWY-5863, PWY-5897, PWY-5898, PWY-5899, PWY-621, PWY-6263, PWY-6317, PWY-6353, PWY-6609, PWY-7208, PWY-7254, PWY-7371, PWY-7373, PWY0-1241, PWY0-1296, PWY0-1297, PWY0-1298, PWY0-1586, PWY0-42, PWY4FS-7, PWY4FS-8, REDCITCYC, SALVADEHYPOX-PWY 
#> 
#> The DESeq2 method obtained 41 statistically significant features.
#> The number of features that are common to other methods is 8 
#> The number of features that are different from other methods is 41 
#> The names of the features that are different from other methods are CENTFERM-PWY, FAO-PWY, FUC-RHAMCAT-PWY, FUCCAT-PWY, HEMESYN2-PWY, HSERMETANA-PWY, NAD-BIOSYNTHESIS-II, P4-PWY, PHOSLIPSYN-PWY, POLYAMINSYN3-PWY, PPGPPMET-PWY, PWY-3781, PWY-4984, PWY-5667, PWY-5747, PWY-5837, PWY-5838, PWY-5840, PWY-5855, PWY-5856, PWY-5857, PWY-5861, PWY-5863, PWY-5897, PWY-5898, PWY-5899, PWY-621, PWY-6263, PWY-6317, PWY-6590, PWY-6708, PWY-7371, PWY0-1241, PWY0-1297, PWY0-1298, PWY0-1319, PWY0-1586, PWY0-42, PWY0-781, REDCITCYC, UBISYN-PWY 
#> 
#> The edgeR method obtained 41 statistically significant features.
#> The number of features that are common to other methods is 8 
#> The number of features that are different from other methods is 41 
#> The names of the features that are different from other methods are ALL-CHORISMATE-PWY, AST-PWY, CENTFERM-PWY, ENTBACSYN-PWY, FAO-PWY, FUCCAT-PWY, GLUCOSE1PMETAB-PWY, GLYCOLYSIS-TCA-GLYOX-BYPASS, NAD-BIOSYNTHESIS-II, P105-PWY, P281-PWY, P562-PWY, PPGPPMET-PWY, PWY-181, PWY-3781, PWY-5088, PWY-5415, PWY-5747, PWY-5837, PWY-5855, PWY-5856, PWY-5857, PWY-5863, PWY-6263, PWY-6590, PWY-6629, PWY-6641, PWY-6708, PWY-6728, PWY-7094, PWY-7254, PWY-7371, PWY-7373, PWY0-1241, PWY0-41, PWY0-42, REDCITCYC, SUCSYN-PWY, TCA-GLYOX-BYPASS, TYRFUMCAT-PWY, UBISYN-PWY 
#> 
#> The Maaslin2 method obtained 30 statistically significant features.
#> The number of features that are common to other methods is 8 
#> The number of features that are different from other methods is 30 
#> The names of the features that are different from other methods are PWY0-1241, PWY-4984, PWY-7373, PWY0-781, PPGPPMET-PWY, PWY-3781, PWY-5747, PWY0-42, PWY0-1586, HEMESYN2-PWY, P4-PWY, PHOSLIPSYN-PWY, PWY-6317, PWY-7254, REDCITCYC, PWY-6641, PWY-6263, PWY-7371, MET-SAM-PWY, ASPASN-PWY, HOMOSER-METSYN-PWY, HSERMETANA-PWY, PWY-5347, PWY-5667, PWY0-1319, ANAGLYCOLYSIS-PWY, PWY-5686, PWY0-1298, PWY-621, P23-PWY 
#> 
#> The number of features that are common to all methods is 8 
#> The names of the features that are common to all methods are PPGPPMET-PWY, PWY-3781, PWY-5747, PWY-6263, PWY-7371, PWY0-1241, PWY0-42, REDCITCYC 
#> The number of features that are obtained by any of the methods is 117 
#> The names of the features that are obtained by any of the methods are 1CMET2-PWY, ANAEROFRUCAT-PWY, ANAGLYCOLYSIS-PWY, ASPASN-PWY, CALVIN-PWY, COA-PWY, DAPLYSINESYN-PWY, DTDPRHAMSYN-PWY, GALACT-GLUCUROCAT-PWY, GALACTUROCAT-PWY, GLCMANNANAUT-PWY, GLUCONEO-PWY, GLUCOSE1PMETAB-PWY, GLUCUROCAT-PWY, GLYCOGENSYNTH-PWY, GLYCOLYSIS, NONOXIPENT-PWY, OANTIGEN-PWY, P161-PWY, PENTOSE-P-PWY, PEPTIDOGLYCANSYN-PWY, PHOSLIPSYN-PWY, POLYAMINSYN3-PWY, PPGPPMET-PWY, PWY-3781, PWY-4984, PWY-5100, PWY-5384, PWY-5484, PWY-5505, PWY-5667, PWY-5686, PWY-5747, PWY-5837, PWY-5838, PWY-5840, PWY-5845, PWY-5850, PWY-5860, PWY-5861, PWY-5862, PWY-5863, PWY-5896, PWY-5897, PWY-5898, PWY-5899, PWY-6151, PWY-621, PWY-6263, PWY-6317, PWY-6385, PWY-6386, PWY-6387, PWY-6471, PWY-6609, PWY-7199, PWY-7208, PWY-7219, PWY-7221, PWY-7229, PWY-7242, PWY-7315, PWY-7371, PWY-7373, PWY0-1241, PWY0-1296, PWY0-1297, PWY0-1298, PWY0-1319, PWY0-1586, PWY0-42, PWY4FS-7, PWY4FS-8, REDCITCYC, SER-GLYSYN-PWY, TRNA-CHARGING-PWY, UDPNAGSYN-PWY, PWY-6353, PWY-7254, SALVADEHYPOX-PWY, CENTFERM-PWY, FAO-PWY, FUC-RHAMCAT-PWY, FUCCAT-PWY, HEMESYN2-PWY, HSERMETANA-PWY, NAD-BIOSYNTHESIS-II, P4-PWY, PWY-5855, PWY-5856, PWY-5857, PWY-6590, PWY-6708, PWY0-781, UBISYN-PWY, ALL-CHORISMATE-PWY, AST-PWY, ENTBACSYN-PWY, GLYCOLYSIS-TCA-GLYOX-BYPASS, P105-PWY, P281-PWY, P562-PWY, PWY-181, PWY-5088, PWY-5415, PWY-6629, PWY-6641, PWY-6728, PWY-7094, PWY0-41, SUCSYN-PWY, TCA-GLYOX-BYPASS, TYRFUMCAT-PWY, MET-SAM-PWY, HOMOSER-METSYN-PWY, PWY-5347, P23-PWY 
# }